+++ packages/gatsby-theme-sso-ui-kit/shortcodes.js
@@ -1,0 +1,1 @@
+// Export only components that must be injected into the MDX provider