+++ packages/gatsby-theme-sso-ui-kit/src/index.ts
@@ -1,0 +1,2 @@
+export { default as LogoutButton } from './components/logout-button';
+export { default as LoginButton } from './components/login-button';